api cleanup
