+++ .storybook/preview-head.html
@@ -1,1 +1,9 @@
-<script>document.documentElement.setAttribute('lang', 'en');</script>
+<!--
+
+  Used to ensure there is a lang set on the html attribute for storybook
+  This is important for accessibility scores
+
+-->
+<script>
+  document.documentElement.setAttribute('lang', 'en');
+</script>
