+++ src/view/drag-drop-context/drag-drop-context.jsx
@@ -3,6 +3,7 @@ import React, { type Node } from 'react';
 import { useMemo } from 'use-memo-one';
 import type { Responders, ContextId, Sensor } from '../../types';
 import ErrorBoundary from '../error-boundary';
+import preset from '../../screen-reader-message-preset';
 import App from './app';
 
 type Props = {|
@@ -12,7 +13,7 @@ type Props = {|
 
   sensors?: Sensor[],
   enableDefaultSensors?: ?boolean,
-  dragHandleDescription?: string,
+  liftInstruction?: string,
 |};
 
 let instanceCount: number = 0;
@@ -22,12 +23,10 @@ export function resetServerContext() {
   instanceCount = 0;
 }
 
-const DEFAULT_DRAG_HANDLE_DESCRIPTION = `Draggable item. Ensure you're screen reader is not in browse mode and then press spacebar to lift.`;
-
 export default function DragDropContext(props: Props) {
   const contextId: ContextId = useMemo(() => `${instanceCount++}`, []);
-  const description =
-    props.dragHandleDescription || DEFAULT_DRAG_HANDLE_DESCRIPTION;
+  const liftInstruction: string =
+    props.liftInstruction || preset.liftInstruction;
 
   // We need the error boundary to be on the outside of App
   // so that it can catch any errors caused by App
@@ -37,7 +36,7 @@ export default function DragDropContext(props: Props) {
         <App
           setOnError={setOnError}
           contextId={contextId}
-          dragHandleDescription={description}
+          liftInstruction={liftInstruction}
           {...props}
         >
           {props.children}
