--- src/view/use-drag-handle-description/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { default } from './use-drag-handle-description';
