+++ src/view/use-lift-instruction/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { default } from './use-lift-instruction';
